requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
django-extensions>=3.2.0
django-debug-toolbar>=4.2.0
whitenoise>=6.5.0
gunicorn>=21.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-widget-tweaks>=1.4.12
openpyxl>=3.1.0
reportlab>=4.0.0